Convert hooks to Typescript
